Update Python version, use separate cltk user, update install_corpora script, use Docker volumes for cltk data
